PR readability: trying the get chunk_filter.c to be recognized as moved.
